Undo changes to loop index in linear morphology.
This might be causing image diffs on MaliG78. The code now
accumulates `delta` via repeated addition instead of multiplication,
avoiding the need to multiply by `i` at all. (Note that `offset`
only contains zeros and ones, so there isn't realistically a
concern about accumulating rounding error.)

Morphology issue in Gold: https://gold.skia.org/diff?grouping=name%3Dfilters-morph-01-f.svg%26source_type%3Dsvg&left=165958d022fa453b8977bfc208e61b25&right=37e489700dc236e0144c5ec1335a6d49

Bug: b/328051548
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/884516
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>

+++ src/effects/imagefilters/SkMorphologyImageFilter.cpp
@@ -25,6 +25,7 @@
 #include "src/core/SkWriteBuffer.h"
 
 #include <algorithm>
+#include <cstdint>
 #include <optional>
 #include <utility>
 
@@ -141,7 +142,7 @@ sk_sp<SkShader> make_linear_morphology(sk_sp<SkShader> input,
     builder.child("child") = std::move(input);
     builder.uniform("offset") = direction == MorphDirection::kX ? SkV2{1.f, 0.f} : SkV2{0.f, 1.f};
     builder.uniform("flip") = (type == MorphType::kDilate) ? 1.f : -1.f;
-    builder.uniform("radius") = (float)radius;
+    builder.uniform("radius") = (int32_t)radius;
 
     return builder.makeShader();
 }
